+++ src/Console/BaseUpdateShell.php
@@ -20,27 +20,46 @@
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
-namespace MeCms\Shell;
+namespace MeCms\Console;
 
 use Cake\Datasource\ConnectionManager;
 use Cake\I18n\Time;
 use MeTools\Console\Shell;
 
+/**
+ * Provided some provides some useful methods for the `UpdateShell` classes
+ */
 class BaseUpdateShell extends Shell
 {
     /**
      * Database connection
-     * @see initialize()
-     * @var resource
+     * @var \Cake\Database\Connection
      */
     protected $connection;
 
     /**
      * Now for MySql
-     * @see initialize()
-     * @var string
+     * @var \Cake\I18n\Time
      */
     protected $now;
+
+    /**
+     * Construct
+     * @param \Cake\Console\ConsoleIo|null $io An io instance
+     * @return void
+     * @uses $connection
+     * @uses $now
+     */
+    public function __construct(\Cake\Console\ConsoleIo $io = null)
+    {
+        parent::__construct($io);
+
+        //Gets database connection
+        $this->connection = ConnectionManager::get('default');
+
+        //Sets now for MySql
+        $this->now = new Time;
+    }
 
     /**
      * Checks if a column exists
@@ -65,21 +84,15 @@ class BaseUpdateShell extends Shell
         $methods = getChildMethods(get_called_class());
 
         return af(array_map(function ($method) {
-            //Filters invalid method names
-            if (!preg_match('/^to([0-9]+)v([0-9]+)v(.+)$/', $method, $matches)) {
-                return false;
+            //Returns array with the name method and the version number
+            if (preg_match('/^to([0-9]+)v([0-9]+)v(.+)$/', $method, $matches)) {
+                return [
+                    'name' => $method,
+                    'version' => $matches[1] . '.' . $matches[2] . '.' . $matches[3],
+                ];
             }
 
-            //Returns array with the name method and the version number
-            return [
-                'name' => $method,
-                'version' => sprintf(
-                    '%s.%s.%s',
-                    $matches[1],
-                    $matches[2],
-                    $matches[3]
-                ),
-            ];
+            return false;
         }, $methods));
     }
 
@@ -96,6 +109,18 @@ class BaseUpdateShell extends Shell
         return array_map(function ($column) {
             return firstValue($column);
         }, $columns);
+    }
+
+    /**
+     * Gets the latest update method.
+     *
+     * Return an array with the name method and the version number.
+     * @return array
+     * @uses _getAllUpdateMethods()
+     */
+    protected function _getLatestUpdateMethod()
+    {
+        return firstValue($this->_getAllUpdateMethods());
     }
 
     /**
@@ -133,9 +158,7 @@ class BaseUpdateShell extends Shell
         $methods = array_reverse($this->_getAllUpdateMethods());
 
         foreach ($methods as $method) {
-            $this->verbose(
-                __d('me_cms', 'Upgrading to {0}', $method['version'])
-            );
+            $this->verbose(__d('me_cms', 'Upgrading to {0}', $method['version']));
 
             //Calls dynamically each method
             $this->{$method['name']}();
@@ -145,18 +168,16 @@ class BaseUpdateShell extends Shell
     /**
      * Performs the latest update available
      * @return void
-     * @uses _getAllUpdateMethods()
+     * @uses _getLatestUpdateMethod()
      */
     public function latest()
     {
-        $method = firstValue($this->_getAllUpdateMethods());
+        list($name, $version) = array_values($this->_getLatestUpdateMethod());
 
-        $this->verbose(
-            __d('me_cms', 'Upgrading to {0}', $method['version'])
-        );
+        $this->verbose(__d('me_cms', 'Upgrading to {0}', $version));
 
         //Calls dynamically the method
-        $this->{$method['name']}();
+        $this->{$name}();
     }
 
     /**
@@ -168,14 +189,8 @@ class BaseUpdateShell extends Shell
     {
         $parser = parent::getOptionParser();
 
-        $parser->addSubcommands([
-            'all' => [
-                'help' => __d('me_cms', 'Performs all available updates'),
-            ],
-            'latest' => [
-                'help' => __d('me_cms', 'Performs the latest update available'),
-            ],
-        ]);
+        $parser->addSubcommand('all', ['help' => __d('me_cms', 'Performs all available updates')]);
+        $parser->addSubcommand('latest', ['help' => __d('me_cms', 'Performs the latest update available')]);
 
         $methods = $this->_getAllUpdateMethods();
 
@@ -187,22 +202,5 @@ class BaseUpdateShell extends Shell
         }
 
         return $parser;
-    }
-
-    /**
-     * Initialize
-     * @return void
-     * @uses $connection
-     * @uses $now
-     */
-    public function initialize()
-    {
-        parent::initialize();
-
-        //Gets database connection
-        $this->connection = ConnectionManager::get('default');
-
-        //Sets now for MySql
-        $this->now = new Time();
     }
 }